+++ sentry-rails/lib/sentry/rails/action_cable/connection.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require_relative 'exception_reporter'
+
+module Sentry
+  module Rails
+    module ActionCable
+      module Connection
+        private
+
+        def handle_open
+          ExceptionReporter.capture(env, transaction_name: self.class.name) { super }
+        end
+      end
+    end
+  end
+end